test: fix tests to use a valid signature
